+++ config/jest/test.config.js
@@ -22,7 +22,6 @@ module.exports = {
     moduleNameMapper: {
         "^@khanacademy/wonder-blocks-(.*)$":
             "<rootDir>/packages/wonder-blocks-$1/src/index.js",
-        "^enzyme$": "<rootDir>/utils/testing/enzyme-shim.js",
     },
     collectCoverageFrom: [
         "packages/**/*.js",