+++ packages/wonder-blocks-link/src/components/__tests__/link.test.js
@@ -1,16 +1,10 @@
 // @flow
 import * as React from "react";
 import {MemoryRouter, Route, Switch} from "react-router-dom";
-import {mount} from "enzyme";
-import "jest-enzyme";
+import {fireEvent, render, screen, waitFor} from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
 
 import Link from "../link.js";
-
-const wait = (delay: number = 0) =>
-    new Promise((resolve, reject) => {
-        // eslint-disable-next-line no-restricted-syntax
-        return setTimeout(resolve, delay);
-    });
 
 describe("Link", () => {
     beforeEach(() => {
@@ -22,12 +16,13 @@ describe("Link", () => {
 
     afterEach(() => {
         window.location.assign.mockClear();
+        jest.clearAllMocks();
     });
 
     describe("client-side navigation", () => {
         test("works for known URLs", () => {
             // Arrange
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link href="/foo">Click me!</Link>
@@ -41,16 +36,16 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {button: 0});
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(wrapper.find("#foo").exists()).toBe(true);
+            expect(screen.getByText("Hello, world!")).toBeInTheDocument();
         });
 
         test("navigation to without route does not render", () => {
             // Arrange
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link href="/unknown">Click me!</Link>
@@ -64,16 +59,16 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {button: 0});
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(wrapper.find("#foo").exists()).toBe(false);
+            expect(screen.queryByText("Hello, world!")).not.toBeInTheDocument();
         });
 
-        test("waits until beforeNav resolves before navigating", async () => {
+        test("waits until beforeNav resolves before navigating", () => {
             // Arrange
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link href="/foo" beforeNav={() => Promise.resolve()}>
@@ -89,20 +84,16 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
-            await wait(0);
-            wrapper.update();
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(wrapper.find("#foo")).toExist();
+            expect(screen.queryByText("Hello, world!")).not.toBeInTheDocument();
         });
 
-        test("doesn't navigate before beforeNav resolves", async () => {
+        test("doesn't navigate before beforeNav resolves", () => {
             // Arrange
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link href="/foo" beforeNav={() => Promise.resolve()}>
@@ -118,18 +109,16 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(wrapper.find("#foo")).not.toExist();
+            expect(screen.queryByText("Hello, world!")).not.toBeInTheDocument();
         });
 
-        test("does not navigate if beforeNav rejects", async () => {
+        test("does not navigate if beforeNav rejects", () => {
             // Arrange
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link href="/foo" beforeNav={() => Promise.reject()}>
@@ -145,21 +134,17 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
-            await wait(0);
-            wrapper.update();
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(wrapper.find("#foo")).not.toExist();
+            expect(screen.queryByText("Hello, world!")).not.toBeInTheDocument();
         });
 
         test("runs safeWithNav if set", async () => {
             // Arrange
             const safeWithNavMock = jest.fn();
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link
@@ -179,21 +164,19 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
-            await wait(0);
-            wrapper.update();
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(safeWithNavMock).toHaveBeenCalled();
+            await waitFor(() => {
+                expect(safeWithNavMock).toHaveBeenCalled();
+            });
         });
 
         test("doesn't run safeWithNav until beforeNav resolves", () => {
             // Arrange
             const safeWithNavMock = jest.fn();
-            const wrapper = mount(
+            render(
                 <MemoryRouter>
                     <div>
                         <Link
@@ -213,10 +196,8 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
             expect(safeWithNavMock).not.toHaveBeenCalled();
@@ -227,7 +208,7 @@ describe("Link", () => {
         test("doesn't redirect if safeWithNav hasn't resolved yet when skipClientNav=true", () => {
             // Arrange
             jest.spyOn(window.location, "assign").mockImplementation(() => {});
-            const wrapper = mount(
+            render(
                 <Link
                     href="/foo"
                     safeWithNav={() => Promise.resolve()}
@@ -238,10 +219,8 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
             expect(window.location.assign).not.toHaveBeenCalled();
@@ -250,7 +229,7 @@ describe("Link", () => {
         test("redirects after safeWithNav resolves when skipClientNav=true", async () => {
             // Arrange
             jest.spyOn(window.location, "assign").mockImplementation(() => {});
-            const wrapper = mount(
+            render(
                 <Link
                     href="/foo"
                     safeWithNav={() => Promise.resolve()}
@@ -261,21 +240,19 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
-            await wait(0);
-            wrapper.update();
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(window.location.assign).toHaveBeenCalledWith("/foo");
+            await waitFor(() => {
+                expect(window.location.assign).toHaveBeenCalledWith("/foo");
+            });
         });
 
         test("redirects after beforeNav and safeWithNav resolve when skipClientNav=true", async () => {
             // Arrange
             jest.spyOn(window.location, "assign").mockImplementation(() => {});
-            const wrapper = mount(
+            render(
                 <Link
                     href="/foo"
                     beforeNav={() => Promise.resolve()}
@@ -287,21 +264,19 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
-            await wait(0);
-            wrapper.update();
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
-            expect(window.location.assign).toHaveBeenCalledWith("/foo");
+            await waitFor(() => {
+                expect(window.location.assign).toHaveBeenCalledWith("/foo");
+            });
         });
 
         test("doesn't redirect before beforeNav resolves when skipClientNav=true", () => {
             // Arrange
             jest.spyOn(window.location, "assign").mockImplementation(() => {});
-            const wrapper = mount(
+            render(
                 <Link
                     href="/foo"
                     beforeNav={() => Promise.resolve()}
@@ -312,10 +287,8 @@ describe("Link", () => {
             );
 
             // Act
-            const linkWrapper = wrapper.find("Link").first();
-            linkWrapper.simulate("click", {
-                button: 0,
-            });
+            const link = screen.getByText("Click me!");
+            userEvent.click(link);
 
             // Assert
             expect(window.location.assign).not.toHaveBeenCalled();
@@ -325,38 +298,38 @@ describe("Link", () => {
     describe("raw events", () => {
         test("onKeyDown", () => {
             // Arrange
-            const keyMock = jest.fn();
-            const wrapper = mount(
-                <Link href="/" onKeyDown={keyMock}>
+            let keyCode;
+            render(
+                <Link href="/" onKeyDown={(e) => (keyCode = e.keyCode)}>
                     Click me!
                 </Link>,
             );
 
             // Act
-            wrapper.find(Link).simulate("keydown", {keyCode: 32});
+            const link = screen.getByText("Click me!");
+            // eslint-disable-next-line testing-library/prefer-user-event
+            fireEvent.keyDown(link, {keyCode: 32});
 
             // Assert
-            expect(keyMock).toHaveBeenCalledWith(
-                expect.objectContaining({keyCode: 32}),
-            );
+            expect(keyCode).toEqual(32);
         });
 
         test("onKeyUp", () => {
             // Arrange
-            const keyMock = jest.fn();
-            const wrapper = mount(
-                <Link href="/" onKeyDown={keyMock}>
+            let keyCode;
+            render(
+                <Link href="/" onKeyUp={(e) => (keyCode = e.keyCode)}>
                     Click me!
                 </Link>,
             );
 
             // Act
-            wrapper.find(Link).simulate("keydown", {keyCode: 32});
+            const link = screen.getByText("Click me!");
+            // eslint-disable-next-line testing-library/prefer-user-event
+            fireEvent.keyUp(link, {keyCode: 32});
 
             // Assert
-            expect(keyMock).toHaveBeenCalledWith(
-                expect.objectContaining({keyCode: 32}),
-            );
+            expect(keyCode).toEqual(32);
         });
     });
 });