+++ packages/wonder-blocks-birthday-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @khanacademy/wonder-blocks-birthday-picker
 
+## 1.2.26
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-dropdown@2.10.2
+
 ## 1.2.25
 
 ### Patch Changes