+++ packages/wonder-blocks-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-button
 
+## 3.0.9
+
+### Patch Changes
+
+-   Updated dependencies [4c682709]
+    -   @khanacademy/wonder-blocks-clickable@2.4.1
+
 ## 3.0.8
 
 ### Patch Changes